some bindings

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/utils/hardware/data/BlinkinInput.kt
@@ -1,3 +1,3 @@
 package org.firstinspires.ftc.teamcode.utils.hardware.data
 
-data class BlinkinInput(val id: Int, val type: BlinkinOptions)
+data class BlinkinInput(val id: Double, val type: BlinkinOptions)
