+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/utils/hardware/requests/BlinkinRequest.kt
@@ -20,7 +20,7 @@ class BlinkinRequest(name: String, hardwareMap: HardwareMap) : ScriptParameters.
         return if (type === BlinkinOptions.GET) {
             blinkin.position
         } else {
-            blinkin.position = Range.clip(id.toDouble(), 0.2525, 0.7475)
+            blinkin.position = Range.clip(id, 0.2525, 0.7475)
             0
         }
     }
